refactor: Extract shared utilities for ScheduleHearing and streamline modal components; enhance date handling and button implementations
